+++ net/ipv6/route.c
@@ -406,7 +406,7 @@ static void ip6_dst_ifdown(struct dst_entry *dst, struct net_device *dev,
 static bool __rt6_check_expired(const struct rt6_info *rt)
 {
 	if (rt->rt6i_flags & RTF_EXPIRES)
-		return time_after(jiffies, rt->dst.expires);
+		return time_after(jiffies, (unsigned long)rt->dst.expires);
 	else
 		return false;
 }
@@ -414,7 +414,7 @@ static bool __rt6_check_expired(const struct rt6_info *rt)
 static bool rt6_check_expired(const struct rt6_info *rt)
 {
 	if (rt->rt6i_flags & RTF_EXPIRES) {
-		if (time_after(jiffies, rt->dst.expires))
+		if (time_after(jiffies, (unsigned long)rt->dst.expires))
 			return true;
 	} else if (rt->dst.from) {
 		return rt6_check_expired((struct rt6_info *) rt->dst.from);
